Waves refactored
