+++ frontend/src/pages/game/Models/Ship/Waves/Wave.jsx
@@ -1,8 +1,8 @@
 import React, { useRef, useEffect } from 'react';
-import gsap from 'gsap';
 import { AdditiveBlending } from 'three';
 import { useSelector } from 'react-redux';
 import { selectTheme } from '../../../../../store/reducers/gameSlice';
+import { waveAnimation } from './animations';
 
 const Wave = ({ id, onComplete }) => {
   const theme = useSelector(selectTheme);
@@ -10,29 +10,7 @@ const Wave = ({ id, onComplete }) => {
   const circleRef = useRef();
 
   useEffect(() => {
-    gsap.fromTo(
-      circleRef.current.scale,
-      { x: 0, y: 0, z: 0 },
-      {
-        x: 0.17,
-        y: 0.17,
-        z: 0.17,
-        duration: 1,
-        onComplete: () => {
-          onComplete?.(id);
-        }
-      }
-    );
-    gsap.to(circleRef.current.material, {
-      opacity: 0,
-      delay: 0.2,
-      duration: 0.6
-    });
-    gsap.to(circleRef.current.material, {
-      emissiveIntensity: 0,
-      delay: 0.2,
-      duration: 0.6
-    });
+    waveAnimation(circleRef, id, onComplete);
   }, []);
 
   return (
